linter: move NodePath to irutil
